+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-## 0.3.3-SNAPSHOT (2019-10-22)
+## 0.3.4-SNAPSHOT (2019-11-05)
 - Added option **autoLoad** in the Delivery annotation so that *DeliverableDispatcher* can still handle the dependency
 injection in the case where the delivery is missing but a corresponding
 repository is present.
 - Added option **condition** in the Delivery annotation to pre-filter loaded data when **autoLoad** is set to true.
+
+## 0.3.3 (2019-10-22)
 - Added **SnappyCompressor**.
 - Added method **persist(persistence: Boolean)** into **Stage** and **Factory** to.
 activate/deactivate output persistence. By default the output persistence is set to *true*.